Moved custom model loader function out of parser.py.
